smart-ml-preprocessor/requirements.txt

kafka-python
pandas
scikit-learn
